requirements.txt
pyserial
msgpack
requests
